Polishing, remove warnings.

+++ CS230_Final/Camera.cpp
@@ -5,8 +5,8 @@
 #define ROTATION_FACTOR 3.14159265f *2.0f / 20.0f
 void Camera::SetSize(long input_right, long input_bottom)
 {
-	right_.x = input_right;
-	up_.y    = input_bottom;
+	right_.x = static_cast<float>(input_right);
+	up_.y    = static_cast<float>(input_bottom);
 }
 
 void Camera::ResizeCamera(short mouse_delta)
